requirements.txt
# Implicit dependencies (optional dependencies of dependencies)
pygments==2.2.0
python-social-auth==0.2.21
elasticsearch==5.5.2
elasticsearch-dsl==5.4.0
sqlparse==0.2.4

# Explicit dependencies (references in code)
Django==1.10.8
django-crispy-forms==1.7.2
django-model-utils==3.1.1
django-munin==0.2.1
python-memcached==1.59
lxml==4.2.1
factory-boy==2.8.1
pygeoip==0.3.2
pillow==5.0.0
gitpython==2.1.9
easy-thumbnails==2.5.0
CairoSVG==2.1.3
beautifulsoup4==4.6.0
django-recaptcha==1.3.1
toml==0.9.4
requests==2.18.4
Pygments==2.2.0

# Api dependencies
djangorestframework==3.6.4
djangorestframework-xml==1.3.0
django-filter==1.1.0
django-oauth-toolkit==1.0.0
drf-extensions==0.3.1
django-rest-swagger==2.1.2
django-cors-middleware==1.3.1
dry-rest-permissions==0.1.10

# Zep 12 dependency
watchdog==0.8.3 # use last non-bc-breaking version
# next versions of this libraries break previous behavior for slug generation from string with single quotes
django-uuslug==1.0.3
python-slugify==1.1.4
